ast: add definitions for `AstNode`, `SpanMap`, `LocalSpanMap`, etc.
This commit adds all of the necessary structures needed to represent and
work with the AST. The `AstNode` is the main structure that represents
the AST. It is a recursive structure that can contain other `AstNode`s.
The `SpanMap` is a structure that maps `Span`s to `AstNode`s. The `Span`
structure is a simple structure that represents a span in the source code.