Introduce `HasSource` trait and implement it for `WorkspaceMembers`
This commit introduces the `HasSource` trait which is used to get the source
and other metadata about an item from a `SourceId`. This is useful for
implementing diagnostics and other features that need to know the source of
an item.

Additionally, with introduce the `TempSourceMap` which is a useful implementer
of `HasSource` for temporary sources, useful for debugging and development purposes.