lexer: implement generic text hunk parsing
Text hunks are areas of the source that are not within the templating language
tags, i.e. not within `{{ ... }}` or `{% ... %}`. This is the most common
source of text in a template, and it is important to be able to parse it
correctly.

For now, we just consume the content until with reach a `{{` or `{%`.

We'll need to find a way of handling the case where we never find an terminating
one. In this scenario, we should then treat it as just plain text (maybe?).