lexer: implement token tree lexing mechanism
Token trees are those that are located within some kind of token context, i.e. a
`{{ ... }}` or `{% ... %}` block, one that has an opening and a closing token to
be valid.

Within those blocks, we essentially maintain a "stack" like mechanism to ensure
that all of the tokens are balanced, and that we can correctly identify the
boundaries of the token tree.

Furthermore, token trees will come into play more during parsing, as we can automatically
infer contexts from the trees, and efficiently parse them (and perform extra error checking)
as we go along.